requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.6.5
pinecone>=3.0.0
sentence-transformers>=2.2.0
requests>=2.31.0
tenacity>=8.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
python-multipart>=0.0.6
pytest>=7.0.0
